pyproject.toml
[tool.poetry]
name = "deployment-manager"
version = "0.0.0"
packages = [
    { include = "deployment_manager" },
]

description = ""
authors = [
    "Miroslav Rehounek <miroslav.rehounek@rossum.ai>",
    "Jan Sporek <jan.sporek@rossum.ai>",
    "Ondrej Polesny <ondrej.polesny@rossum.ai>",
]
license = "MIT"
readme = "README.md"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
pattern = "v(?P<base>\\d+\\.\\d+\\.\\d+)"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.25.2"
click = "^8.1.7"
pyyaml = "^6.0.1"
anyio = "^4.2.0"
rich = "^13.7.0"
jmespath = "^1.0.1"
ruamel-yaml = "^0.18.6"
pydantic = "^2.9.2"
rossum-api = "^3.13.2"
questionary = "^2.0.1"
boto3 = "^1.37.30"
botocore = "^1.37.30"

# Moved from `rossum_api` dependencies:
tenacity = "^8.2.3"
datamodel-code-generator = "^0.21.2"
fastapi = "^0.100.0"
uvicorn = "^0.20.0"
aiofiles = "^24.1.0"
dacite = "^1.9.2"
diagrams = "^0.24.4"
langchain = {extras = ["aws"], version = "^0.3.25"}
langchain-aws = "^0.2.23"

[tool.poetry.scripts]
prd2 = "deployment_manager.main:main"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.7"
ruff = "^0.5.0"
pre-commit = "^4.0.1"
pytest-dotenv = "^0.5.2"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.ruff]
exclude = ["dummy_project", "dist"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["deployment_manager"]

[tool.pytest]
env_files = [".env"]

[tool.coverage.run]
source = ["deployment_manager"]
branch = true
# Use relative paths in coverage.xml so GitLab can map them to MR file paths.
relative_files = true

[tool.coverage.report]
# Fail CI if coverage drops below this (tweak as coverage stabilises).
# fail_under = 80
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

[tool.coverage.xml]
output = "coverage.xml"
